+++ modAionImpl/src/org/aion/zero/impl/ChainStatistics.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.manager;
+package org.aion.zero.impl;
 
 import java.util.LinkedList;
 import java.util.List;
