+++ modAionImpl/src/org/aion/zero/impl/blockchain/AionPendingStateImpl.java
@@ -41,7 +41,6 @@ import org.aion.mcf.core.AccountState;
 import org.aion.mcf.db.IBlockStoreBase;
 import org.aion.mcf.db.Repository;
 import org.aion.mcf.db.RepositoryCache;
-import org.aion.mcf.types.PendingTransactionState;
 import org.aion.zero.impl.db.TransactionStore;
 import org.aion.mcf.valid.TransactionTypeRule;
 import org.aion.mcf.valid.TxNrgRule;
