+++ modAionImpl/test/org/aion/zero/impl/db/AionRepositoryCacheTest.java
@@ -13,7 +13,7 @@ import org.aion.mcf.config.CfgPrune;
 import org.aion.mcf.db.ContractDetails;
 import org.aion.mcf.db.PruneConfig;
 import org.aion.mcf.db.RepositoryConfig;
-import org.aion.mcf.vm.DataWord;
+import org.aion.mcf.vm.types.DataWord;
 import org.aion.mcf.vm.types.DataWordImpl;
 import org.aion.mcf.vm.types.DoubleDataWord;
 import org.aion.types.AionAddress;
