+++ modMcf/src/module-info.java
@@ -25,14 +25,11 @@ module aion.mcf {
     exports org.aion.mcf.db;
     exports org.aion.mcf.db.exception;
     exports org.aion.mcf.ds;
-    exports org.aion.mcf.evt;
-    exports org.aion.mcf.manager;
     exports org.aion.mcf.mine;
     exports org.aion.mcf.serial;
     exports org.aion.mcf.trie;
     exports org.aion.mcf.tx;
     exports org.aion.mcf.types;
     exports org.aion.mcf.valid;
-    exports org.aion.mcf.vm;
     exports org.aion.mcf.vm.types;
 }
