--- modMcf/src/org/aion/mcf/evt/EvtCb.java
@@ -1,6 +1,0 @@
-package org.aion.mcf.evt;
-
-@FunctionalInterface
-public interface EvtCb<T extends EvtData> {
-    void call(T t);
-}
