--- modMcf/src/org/aion/mcf/evt/EvtData.java
@@ -1,4 +1,0 @@
-package org.aion.mcf.evt;
-
-/** @author jin */
-public interface EvtData {}
