+++ modMcf/src/org/aion/mcf/vm/types/DataWordImpl.java
@@ -3,7 +3,7 @@ package org.aion.mcf.vm.types;
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-import org.aion.mcf.vm.DataWord;
+
 import org.aion.util.bytes.ByteUtil;
 import org.aion.util.conversions.Hex;
 import org.aion.util.types.ByteArrayWrapper;
